requirements.txt
PyMuPDF
pdfplumber
thefuzz[speedup]
camelot-py[cv]
python-dotenv==1.0.1
pydantic==2.7.4
fastapi==0.111.0
uvicorn[standard]==0.30.1
tqdm==4.66.4
faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0
torch>=2.2.0
python-docx==1.1.0
PyPDF2==3.0.1
langchain==0.2.8
tiktoken==0.7.0
python-box==7.3.2
PyYAML==6.0.2